refactor: remove old Snap clients
