+++ app/components/Views/MultichainAccounts/WalletDetails/BaseWalletDetails/index.test.tsx
@@ -116,12 +116,6 @@ jest.mock(
   }),
 );
 
-jest.mock('../../../../../core/SnapKeyring/MultichainWalletSnapClient', () =>
-  jest.requireActual(
-    '../../../../../core/SnapKeyring/MultichainWalletSnapClient',
-  ),
-);
-
 const mockGetInternalAccountsFromWallet =
   getInternalAccountsFromWallet as jest.Mock;
 const mockUseWalletBalances = useWalletBalances as jest.Mock;
